Cleaner, Faster and Better
The code has been cleaned up and greatly optimised so that it runs
faster with less iterations and is more precise in it's actions. One big
addition is the translation function for the camera so that you pass in
it's translation relative to camera space and it makes the same shift in
world space without you needing to account for the camera's orientation.
much easier to use.