Revisions
Reviewing code now that some difficulties and impracticalities have been
found when implementing the graphics engine. Objects have been reworked
and the ability to use textures should be implemented soon. Some
optimisation of code as well such as the automated production of
TransformationMatrix's as well as some new usefull methods in the
Point3D and FPoint3D classes.