Progress Me Thinks.
After working on fixing the rendering code i optimised the code further
and fixed a few subtle bugs. I believe the rendering code is working
again but funnily enough I'm not entirely sure. Any bugs seem to be in
the scaling for perspective however it could simply be the wireframe
rendering confusing me.